+++ app/src/main/java/com/example/todoproject/viewmodel/ItemAddViewModel.kt
@@ -1,8 +1,11 @@
-package com.example.todoproject
+package com.example.todoproject.viewmodel
 
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
+import com.example.todoproject.model.ItemDatabase
+import com.example.todoproject.model.TaskRepository
+import com.example.todoproject.model.Item
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
@@ -18,7 +21,10 @@ class ItemAddViewModel(application: Application) : AndroidViewModel(application)
     private val repository: TaskRepository
 
     init {
-        val itemDao = ItemDatabase.getInstance(application, scope).itemDao()
+        val itemDao = ItemDatabase.getInstance(
+            application,
+            scope
+        ).itemDao()
         repository = TaskRepository(itemDao)
         list = repository.list
     }
