+++ app/src/main/res/layout/fragment_item.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="itemAddViewModel"
-            type="com.example.todoproject.ItemAddViewModel" />
+            type="com.example.todoproject.viewmodel.ItemAddViewModel" />
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
